+++ src/is_nothing.ts
@@ -1,6 +1,5 @@
 import type { Nothing } from './utilities';
 
 /** Checks if value is `null` or `undefined`. */
-export const isNothing = <T>(value: T | Nothing): value is Nothing => {
-  return value === null || value === undefined;
-};
+export const isNothing = <T>(value: T | Nothing): value is Nothing =>
+  value === null || value === undefined;