+++ src/is_promise.ts
@@ -1,4 +1,3 @@
 /** Checks if value is `Promise`. */
-export const isPromise = <T>(value: any): value is Promise<T> => {
-  return value instanceof Promise;
-};
+export const isPromise = <T>(value: any): value is Promise<T> =>
+  value instanceof Promise;