+++ test/reviver.spec.js
@@ -1,4 +1,11 @@
-import { isEither, isList, isMaybe, reviver } from '../build';
+import {
+  isList,
+  isMaybe,
+  isTuple,
+  reviver,
+  isEither,
+  isContainer,
+} from '../build';
 
 describe('reviver', () => {
   test('should skip values that are not Maybe, Either and List', () => {
@@ -27,5 +34,15 @@ describe('reviver', () => {
       '{"type":"Either","value":{"type":"Error","value":"Message!"}}';
     expect(isEither(JSON.parse(json, reviver))).toBe(true);
     expect(JSON.parse(json, reviver).isLeft()).toBe(true);
+  });
+
+  test('should create Container instance', () => {
+    const json = '{"type":"Container","value":5}';
+    expect(isContainer(JSON.parse(json, reviver))).toBe(true);
+  });
+
+  test('should create Tuple instance', () => {
+    const json = '{"type":"Tuple","value":[5]}';
+    expect(isTuple(JSON.parse(json, reviver))).toBe(true);
   });
 });