+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Default error type to `Task` class, `task`, `done` and `fail` functions.
 - Add `Lazy` monad and `isLazy` function.
+- `Tuple` type and `isTuple` function.
+- `First`, `Widen`, `Shift`, `Pop`, `Head`, `Transform`, `Position` utility types.
 
 ### Changed
 
@@ -11,6 +13,10 @@
 - Change signature of `sequence` function to accept more than one argument.
 - Rename `Applicative` to `Apply` according to [Fantasy Land specification](https://github.com/fantasyland/fantasy-land#apply)
 - Rename `Wrapper` type to `Container` and `isWrapper` function to `isContainer`.
+- Make `Container` serializable.
+- `tuple` function constructs now `Tuple` object.
+- `reviver` can now parse `Container` and `Tuple` from JSON string.
+- Move `IterableFunction` to _types.ts_ and rename it to `IterableIteratorFunction`.
 
 ### Fixed